+++ README.md
@@ -1,0 +1,40 @@
+# Instrucciones para ejecutar este proyecto
+
+- Clonar el proyecto y cambiar de rama
+```bash
+git clone https://github.com/coder-live-class/django-coderhouse-project.git
+
+cd django-coderhouse-project
+
+git checkout class_20
+
+```
+
+- Crear y activar entorno virtual (Windows)
+```bash
+C:\>python -m venv c:\ruta\al\entorno\virtual
+C:\>c:\ruta\al\entorno\virtual\scripts\activate.bat
+```
+
+- Crear y activar entorno virtual (Linux)
+```bash
+python3 -m venv venv
+source venv/bin/activate
+```
+
+- Crear base de datos con los Modelos (hacer migraciones y migrar)
+```bash
+python manage.py makemigrations app_coder
+
+python manage.py migrate
+```
+
+- Crear super-usuario
+```bash
+python manage.py createsuperuser
+```
+
+- Ejecutar proyecto
+```bash
+python manage.py runserver
+```
