+++ app_coder/admin.py
@@ -1,0 +1,11 @@
+from django.contrib import admin
+
+from app_coder.models import Course, Student, Profesor, Homework
+
+admin.site.register(Course)
+
+admin.site.register(Student)
+
+admin.site.register(Profesor)
+
+admin.site.register(Homework)
