+++ app_coder/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CoderCourseConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'app_coder'
