+++ app_coder/templates/app_coder/courses.html
@@ -1,0 +1,16 @@
+{% extends "app_coder/index.html" %}
+{% load static %}
+{% block contenidoQueCambia %}
+
+<h2>Lista de Cursos</h2>
+<hr></hr>
+<ul>
+    {% for course in courses %}
+        <li>
+            Curso: {{course.name}}
+            -- Camada: {{course.code}}
+        </li>
+    {% endfor %}
+</ul>
+
+{% endblock %}
