+++ app_coder/templates/app_coder/post_course.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <h1>Post courses 2: Using templates of the app "app_coder" and "from django.shortcuts import render"</h1>
+    <hr></hr>
+    <h2>Course: {{course.name}}</h2>
+    <h3>Code: {{course.code}}</h3>
+</body>
+</html>
