+++ app_coder/templates/app_coder/profesors.html
@@ -1,0 +1,17 @@
+{% extends "app_coder/index.html" %}
+{% load static %}
+{% block contenidoQueCambia %}
+
+<h2>Lista de Profesores</h2>
+<hr></hr>
+<ul>
+    {% for profesor in profesors %}
+        <li>
+            Profesor: {{profesor.name}} {{profesor.last_name}}
+            -- e-mail: {{profesor.email}}
+            -- Profesión: {{profesor.profession}}
+        </li>
+    {% endfor %}
+</ul>
+
+{% endblock %}
