+++ app_coder/templates/app_coder/students.html
@@ -1,0 +1,16 @@
+{% extends "app_coder/index.html" %}
+{% load static %}
+{% block contenidoQueCambia %}
+
+<h2>Lista de Estudiantes</h2>
+<hr></hr>
+<ul>
+    {% for student in students %}
+        <li>
+            Estudiante: {{student.name}} {{student.last_name}}
+            -- e-mail: {{student.email}}
+        </li>
+    {% endfor %}
+</ul>
+
+{% endblock %}
