--- coder_course/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CoderCourseConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'coder_course'
