--- coder_course/models.py
@@ -1,25 +1,0 @@
-from django.db import models
-
-
-class Course(models.Model):
-    name = models.CharField(max_length=40)
-    code = models.IntegerField()
-
-
-class Student(models.Model):
-    name = models.CharField(max_length=40)
-    last_name = models.CharField(max_length=40)
-    email = models.EmailField()
-
-
-class Profesor(models.Model):
-    name = models.CharField(max_length=40)
-    last_name = models.CharField(max_length=40)
-    email = models.EmailField()
-    profession = models.CharField(max_length=40)
-
-
-class Homeworks(models.Model):
-    name = models.CharField(max_length=40)
-    due_date = models.DateField()
-    is_delivered = models.BooleanField()
