--- coder_course/templates/coder_course/get_course.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <h1>Get courses 2: Using templates of the app "coder_course" and "from django.shortcuts import render"</h1>
-    <hr></hr>
-    <h2>Course:  {{course.name}}</h2>
-    <h3>ID: {{course.id}}</h3>
-</body>
-</html>
