--- live_class/templates/all_courses.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <h1>All courses 1: Using project templates and "from django.template import loader"</h1>
-    <hr></hr>
-    <h1>The list of the available courses in DB</h1>
-    <hr></hr>
-    <ul>
-        {% for course in courses %}
-            <li>Course {{course.name}}  --   code: {{course.code}}</li>
-        {% endfor %}
-    </ul>
-</body>
-</html>
