--- live_class/templates/get_course.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <h1>Get courses 1: Using project templates and "from django.template import loader"</h1>
-    <hr></hr>
-    <h2>Course:  {{course.name}}</h2>
-    <h3>ID: {{course.id}}</h3>
-</body>
-</html>
